Updated Max Hosts Check
Fixed the check for both Max Range being a number and being greater than Min range

The original primitive version had separate while loops for check- You could bypass the "being the number" by entering a number lower than Min Range.  Then, you could enter a non-number and break the code.

WELL NOT ANYMORE!

The WHile loop has been combined into one and has specific checks for whether or not you are entering a number or a number lower than Min Range, and prompts you correctly on which it is. And you cannot bypass it anymore!